modal removed and file renamed
